warings and dead code from gcc on windows
[git-p4: depot-paths = "//depot/qt/main/": change = 161836]

+++ src/gui/styles/qwindowsstyle.cpp
@@ -1287,8 +1287,8 @@ void QWindowsStyle::drawPrimitive(PrimitiveElement pe, const QStyleOption *opt, 
                 QMatrix wm = p->deviceMatrix();
                 // We cannot use the native function if we have xforms
                 if (wm.m11() == 1 && wm.m22() == 1 && wm.m12() == 0 && wm.m21() == 0) {
-                    RECT rect = { opt->rect.left() + wm.dx(), opt->rect.top() + wm.dy(),
-                                  opt->rect.right() + 1 + wm.dx(), opt->rect.bottom() + 1 + wm.dy() };
+                    RECT rect = { LONG(opt->rect.left() + wm.dx()), LONG(opt->rect.top() + wm.dy()),
+                                  LONG(opt->rect.right() + 1 + wm.dx()), LONG(opt->rect.bottom() + 1 + wm.dy()) };
                     // Force update the HDC before we use it.
                     p->paintEngine()->syncState();
                     HDC hdc = p->device()->getDC();
